# docs/swagger.yaml
basePath: /
host: localhost:8080
info:
  contact: {}
  description: Production-ready multi-tenant payment notification engine for MTN MoMo.
  title: PayKit — Mobile Money Webhook Engine
  version: "1.0"
paths:
...
  /transactions:
    get:
      description: Returns paginated transactions for the authenticated merchant.
        Filter by status or look up a specific order using external_id.
      parameters:
      - description: Look up by order ID (e.g. order_987)
        in: query
        name: external_id
        type: string
      - description: 'Filter by status: SUCCESSFUL, FAILED, PENDING'
        in: query
        name: status
        type: string
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Results per page (default: 20)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              count:
                type: integer
              data:
                type: array
              limit:
                type: integer
              page:
                type: integer
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: List and filter transactions
      tags:
      - Transactions
  /transactions/{id}:
    get:
      description: Returns a single transaction scoped to the authenticated merchant.
        The paid field gives a simple true/false answer.
      parameters:
      - description: Provider transaction ID (e.g. TX-001)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              amount:
                type: string
              currency:
                type: string
              external_id:
                type: string
              paid:
                type: boolean
              provider_tx_id:
                type: string
              received_at:
                type: string
              status:
                type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get transaction by provider ID
      tags:
      - Transactions
...
securityDefinitions:
  BearerAuth:
    description: 'Enter your API key as: Bearer pk_live_xxx'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"